# test/encoding/check-generated: parallel testing by type