@rosinL 2024-03-04
The `omini*` arm machines have poor single core, but many cores, the unittest `check-generated.sh` runs too slow. Parallel testing by type to use all the cores. After change, the unittest `check-generated.sh` time reduced from 7185.87 sec to 1220.25 sec, the Total Test time reduced from 7354.85 sec to 3596.27 sec, 2 times faster than before.
Before change on omini005:
```
284/290 Test #257: unittest-omap-manager .....................   Passed  1433.52 sec
285/290 Test #264: unittest-fltree-onode-manager .............   Passed  1511.33 sec
286/290 Test  #52: unittest_bufferlist .......................   Passed  2285.95 sec
287/290 Test  #32: run-rbd-unit-tests-61.sh ..................   Passed  2800.96 sec
288/290 Test #251: unittest-transaction-manager ..............   Passed  2658.07 sec
289/290 Test #148: readable.sh ...............................   Passed  3161.80 sec
290/290 Test #147: check-generated.sh ........................   Passed  7185.87 sec

99% tests passed, 3 tests failed out of 290

Total Test time (real) = 7354.85 sec
```
After change on omin005:
```
275/289 Test #147: check-generated.sh ........................   Passed  1220.25 sec
276/289 Test #257: unittest-seastore .........................   Passed  505.88 sec
277/289 Test #251: unittest-btree-lba-manager ................   Passed  706.17 sec
278/289 Test #254: unittest-object-data-handler ..............   Passed  802.35 sec
279/289 Test   #4: run-tox-mgr ...............................   Passed  1797.32 sec
        Start   3: teardown-venv-for-mgr
280/289 Test   #3: teardown-venv-for-mgr .....................   Passed    2.03 sec
281/289 Test  #34: run-rbd-unit-tests-127.sh .................   Passed  2084.72 sec
282/289 Test #167: unittest_erasure_code_shec_all ............   Passed  1916.94 sec
283/289 Test #256: unittest-omap-manager .....................   Passed  1477.49 sec
284/289 Test  #31: run-rbd-unit-tests-1.sh ...................   Passed  2349.84 sec
285/289 Test #263: unittest-fltree-onode-manager .............   Passed  1536.68 sec
286/289 Test  #52: unittest_bufferlist .......................   Passed  2558.27 sec
287/289 Test  #32: run-rbd-unit-tests-61.sh ..................   Passed  3036.44 sec
288/289 Test #250: unittest-transaction-manager ..............   Passed  2829.04 sec
289/289 Test #148: readable.sh ...............................   Passed  3431.07 sec

99% tests passed, 3 tests failed out of 289

Total Test time (real) = 3596.27 sec
```